src/rexode_cli/tools/ai_data_processing.py
from PIL import Image
import pytesseract

def summarize_text(text):
    """Summarizes a long text."""
    # This is a placeholder and would require integration with an LLM
    return f"Summarized text: {text[:100]}..."

def translate_text(text, target_language):
    """Translates text from one language to another."""
    # This is a placeholder and would require integration with a translation API or LLM
    return f"Translated '{text}' to {target_language}"

def extract_text_from_image(image_path):
    """Extracts text from an image."""
    try:
        return pytesseract.image_to_string(Image.open(image_path))
    except Exception as e:
        return f"Error extracting text from image: {e}"

def generate_code(prompt, language):
    """Generates code in a specified language."""
    # This is a placeholder and would require integration with an LLM
    return f"Generated {language} code for prompt: {prompt}"

def analyze_data(data_path):
    """Analyzes a dataset and provides insights."""
    # This is a placeholder and would require integration with a data analysis library (e.g., pandas, numpy)
    return f"Analyzed data at {data_path}"
