src/rexode_cli/tools/developer_power_tools.py
import git

def git_clone(repository_url, destination_path):
    """Clones a Git repository."""
    try:
        git.Repo.clone_from(repository_url, destination_path)
        return f"Successfully cloned repository from {repository_url} to {destination_path}"
    except Exception as e:
        return f"Error cloning repository: {e}"

def git_commit_push(commit_message, repository_path):
    """Commits and pushes changes to a Git repository."""
    try:
        repo = git.Repo(repository_path)
        repo.git.add(A=True)
        repo.index.commit(commit_message)
        origin = repo.remote(name='origin')
        origin.push()
        return f"Successfully committed and pushed changes to repository at {repository_path}"
    except Exception as e:
        return f"Error committing and pushing changes: {e}"

def git_pull(repository_path):
    """Pulls changes from a Git repository."""
    try:
        repo = git.Repo(repository_path)
        origin = repo.remote(name='origin')
        origin.pull()
        return f"Successfully pulled changes from repository at {repository_path}"
    except Exception as e:
        return f"Error pulling changes: {e}"

def run_tests(project_path):
    """Runs tests for a project."""
    # This is a placeholder and would require a testing framework (e.g., pytest, unittest)
    return f"Ran tests for project at {project_path}"

def build_project(project_path):
    """Builds a project."""
    # This is a placeholder and would require a build tool (e.g., make, webpack)
    return f"Built project at {project_path}"

def deploy_app(app_path):
    """Deploys an application."""
    # This is a placeholder and would require a deployment script or tool (e.g., Docker, Kubernetes)
    return f"Deployed application at {app_path}"

def manage_docker(command):
    """Manages Docker containers."""
    # This is a placeholder and would require the Docker SDK for Python
    return f"Executed Docker command: {command}"

def database_query(database_url, query):
    """Queries a database."""
    # This is a placeholder and would require a database driver (e.g., psycopg2, mysql-connector-python)
    return f"Executed query on database at {database_url}"

def data_pipeline_run(pipeline_name):
    """Runs a data pipeline."""
    # This is a placeholder and would require a data pipeline framework (e.g., Airflow, Prefect)
    return f"Ran data pipeline: {pipeline_name}"

def generate_api(specification_path):
    """Generates an API from a specification."""
    # This is a placeholder and would require an API generation tool (e.g., FastAPI, Connexion)
    return f"Generated API from specification at {specification_path}"

def code_review(code_path):
    """Performs a code review."""
    # This is a placeholder and would require integration with a code analysis tool or LLM
    return f"Performed code review for code at {code_path}"
